FBV/requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2025.6
Django==4.2.23
django-crispy-forms==2.4
djangorestframework==3.15.2
drf-yasg==1.21.10
gunicorn==23.0.0
inflection==0.5.1
packaging==25.0
psycopg2-binary==2.9.10
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
